+++ api/pacman-api-admin/src/main/java/com/tmobile/pacman/api/admin/repository/service/AccountsService.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  ******************************************************************************/
-package com.tmobile.pacman.api.admin.repository.service.accounts;
+package com.tmobile.pacman.api.admin.repository.service;
 
 import com.tmobile.pacman.api.admin.domain.AccountList;
 import com.tmobile.pacman.api.admin.domain.AccountValidationResponse;
